resolve json serde with non string key and unify object id call
