+++ Cargo.toml
@@ -162,9 +162,10 @@ rayon = "1.5.2"
 rand = "0.8.5"
 rand_core = { version = "0.6.3", default-features = false }
 schemars = { version = "0.8.10", features = ["either"] }
-serde = { version = "1.0.137", features = ["derive", "rc"] }
+serde = { version = "1.0.144", features = ["derive", "rc"] }
 serde_bytes = "0.11.6"
-serde_json = { version = "1.0.111", features = ["preserve_order", "arbitrary_precision"] }
+#serde_json = { version = "1.0.111", features = ["preserve_order", "arbitrary_precision"] }
+serde_json = { version = "1.0.111", features = ["preserve_order"] }
 serde_yaml = "0.9"
 serde_repr = "0.1"
 serde-name = "0.2"
@@ -245,6 +246,7 @@ criterion = { version = "0.5.1", features = [
 ] }
 #criterion-cpu-time = "0.1.0"
 pprof = { version = "0.13.0", features = ["flamegraph", "criterion", "cpp", "frame-pointer"] }
+serde_json_any_key = "2.0.0"
 
 
 celestia-rpc = { git = "https://github.com/eigerco/celestia-node-rs.git", rev = "129272e8d926b4c7badf27a26dea915323dd6489" }
