+++ crates/rooch-indexer/src/tests/test_indexer.rs
@@ -14,7 +14,7 @@ use move_core_types::language_storage::StructTag;
 use move_core_types::vm_status::KeptVMStatus;
 use moveos_types::h256::H256;
 use moveos_types::move_types::{random_struct_tag, random_type_tag};
-use moveos_types::moveos_std::object::ObjectID;
+use moveos_types::moveos_std::object_id::ObjectID;
 use moveos_types::moveos_std::tx_context::TxContext;
 use moveos_types::state::{MoveStructType, SplitStateChangeSet};
 use moveos_types::transaction::{TransactionExecutionInfo, VerifiedMoveOSTransaction};
@@ -273,7 +273,6 @@ fn test_state_store() -> Result<()> {
 
     let coin_info_type =
         StructTag::from_str(format_struct_tag(CoinInfo::<GasCoin>::struct_tag()).as_str())?;
-    // println!("")
     let filter = GlobalStateFilter::ObjectType(coin_info_type);
     let query_global_states =
         indexer_reader.query_global_states_with_filter(filter, None, 1, true)?;
