+++ crates/rooch-rpc-api/src/jsonrpc_types/move_types.rs
@@ -12,12 +12,13 @@ use move_core_types::{
 use move_resource_viewer::{AnnotatedMoveStruct, AnnotatedMoveValue};
 use moveos_types::move_std::string::MoveString;
 use moveos_types::move_types::parse_module_id;
+use moveos_types::moveos_std::object_id::ObjectID;
 use moveos_types::moveos_std::type_info::TypeInfo;
 use moveos_types::transaction::MoveAction;
 use moveos_types::{
     access_path::AccessPath,
     move_types::FunctionId,
-    moveos_std::object::{AnnotatedObject, ObjectID},
+    moveos_std::object::AnnotatedObject,
     transaction::{FunctionCall, ScriptCall},
 };
 use moveos_types::{move_std::ascii::MoveAsciiString, state::MoveStructType};
