+++ crates/rooch-types/src/framework/coin.rs
@@ -7,7 +7,7 @@ use move_core_types::u256::U256;
 use move_core_types::{account_address::AccountAddress, ident_str, identifier::IdentStr};
 use moveos_types::module_binding::{ModuleBinding, MoveFunctionCaller};
 use moveos_types::move_std::string::MoveString;
-use moveos_types::moveos_std::object::{self, ObjectID};
+use moveos_types::moveos_std::object_id::{self, ObjectID};
 use moveos_types::state::{MoveState, MoveStructState, MoveStructType, PlaceholderStruct};
 use serde::{Deserialize, Serialize};
 
@@ -26,7 +26,7 @@ impl<'a> CoinModule<'a> {
     pub fn coin_info_id(coin_type: StructTag) -> ObjectID {
         let coin_info_struct_tag =
             CoinInfo::<PlaceholderStruct>::struct_tag_with_coin_type(coin_type);
-        object::named_object_id(&coin_info_struct_tag)
+        object_id::named_object_id(&coin_info_struct_tag)
     }
 }
