+++ moveos/moveos-store/src/tests/test_state_store.rs
@@ -9,7 +9,7 @@ use moveos_types::h256::H256;
 use moveos_types::move_std::string::MoveString;
 use moveos_types::move_types::random_type_tag;
 use moveos_types::moveos_std::context;
-use moveos_types::moveos_std::object::{NamedTableID, ObjectID};
+use moveos_types::moveos_std::object_id::{NamedTableID, ObjectID};
 use moveos_types::state::{KeyState, MoveState, MoveType, State, StateChangeSet, TableChange};
 use rand::{thread_rng, Rng};
 use smt::NodeStore;
