+++ moveos/moveos-types/src/moveos_std/account_storage.rs
@@ -1,7 +1,7 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
-use super::object::{NamedTableID, ObjectID};
+use super::object_id::{NamedTableID, ObjectID};
 use crate::{
     addresses::MOVEOS_STD_ADDRESS,
     state::{MoveStructState, MoveStructType},
