+++ moveos/moveos-types/src/moveos_std/event.rs
@@ -6,7 +6,7 @@
 
 use crate::addresses::MOVEOS_STD_ADDRESS;
 use crate::h256;
-use crate::moveos_std::object::ObjectID;
+use crate::moveos_std::object_id::ObjectID;
 use crate::state::MoveStructType;
 use anyhow::{ensure, Error, Result};
 use move_core_types::account_address::AccountAddress;
@@ -24,7 +24,7 @@ use crate::module_binding::{ModuleBinding, MoveFunctionCaller};
 use crate::moveos_std::tx_context::TxContext;
 use crate::transaction::FunctionCall;
 
-use super::object;
+use super::object_id;
 
 /// Rust bindings for MoveosStd event module
 pub struct EventModule<'a> {
@@ -281,7 +281,7 @@ impl EventHandle {
     }
 
     pub fn derive_event_handle_id(event_handle_type: &StructTag) -> ObjectID {
-        object::named_object_id(event_handle_type)
+        object_id::named_object_id(event_handle_type)
     }
 }
